Remove dependence on parent for Header and Record.
